fixed build error: changed TexFormatTuple::format type to uint32

+++ src/kani/texture/kaniTexFormat.cpp
@@ -23,7 +23,7 @@ namespace kani { namespace texture {
 	s_TexFormatTupleMap.push_back	\
 	( TexFormatTuple(BOOST_PP_STRINGIZE(SEXTUPLE_1ST(elem)), \
 		(PixelType)SEXTUPLE_2ND(elem), \
-		BOOST_PP_CAT(data, SEXTUPLE_1ST(elem)),\
+		(uint32)BOOST_PP_CAT(data, SEXTUPLE_1ST(elem)),\
 		SEXTUPLE_3RD(elem),	\
 		SEXTUPLE_4TH(elem),	\
 		(uint32)SEXTUPLE_5TH(elem),	\
